Removed outdated information from the README file
I removed what was updated, like the link to packetbeat.com. I also
removed the list of dashbaords because that will always be hard to
maintain the repository.

I didn't update the screenshots with the PR. In general, I think
having no screenshots is worse than having slightly out of date
screenshots, so I'm not too worried about that part lagging behind
occasionally.