Update search with type: process
Also, it removes 2 widgets from the Topbeat dashboard as they are not very useful.
It saves the latest version of the Topbeat index template in Kibana, after adding beat.hostname and beat.name.
By uploading this index pattern to Kibana, the user doesn't need to do a refresh in Kibana to force an update
for all the fields.
Generate Proc stats with Kibana 2.0 as it adds require_field_match:false